pyproject.toml
[project]
name = "QaAgent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.15",
    "asyncclick>=8.1.8",
    "chromadb>=1.0.17",
    "google-adk>=1.7.0",
    "mcp[cli]>=1.12.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pymupdf",
    "langchain",
    "sentence-transformers>=5.1.0",
]

[tool.uv.workspace]
members = [
    "frontend",
]
